requirements.txt
django
gunicorn
psycopg2
south
django-debug-toolbar
bpython
django-mptt
jsonfield
PIL
lxml
easy-thumbnails
django-haystack
pysolr